better handling of unset image fill property.
Unset value is now 0x0 and this is handled as invalid, with an error message.

1x1 is a valid fill, but it is very slow and often system hangs while
it scale the whole thing... usually nobody want it at 1x1, we just end
using that for unset values. With unset value at 0x0 it will not
happen and we'll know when we forgot to do so!.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/evas@38071 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33